binance-wallet/requirements.txt
python-binance
pymongo
python-dotenv
